getting up to date with DShield-ISC

+++ bin/pifwparser.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python
 
 import sys
+from sys import argv
 import os
 import re
 import json
@@ -9,9 +10,9 @@ from time import mktime
 from datetime import datetime
 from DShield import DshieldSubmit
 
-logfile = "/var/log/ufw.log"
-pidfile = "/var/run/ufwparser.pid"
-lastcount = ".lastufw"
+logfile = "/var/log/dshield.log"
+pidfile = "/var/run/pifwparser.pid"
+lastcount = ".lastpifw"
 config = "/etc/dshield.ini"
 startdate = 0
 now = datetime.utcnow()
@@ -20,30 +21,31 @@ fieldmap = {'SRC': 'sip', 'DST': 'dip', 'PROTO': 'proto', 'TYPE': 'sport',
 protomap = {'UDP': 17, 'TCP': 6, 'ICMP': 1, 'ICMPv6': 58}
 tcpflagmap = {'CWR': '1', 'ECE': '2', 'URG': 'U', 'ACK': 'A', 'PSH': 'P', 'RST': 'R', 'SYN': 'S', 'FIN': 'F'}
 d = DshieldSubmit('')
-
+args = d.getopts(argv)
+debug = 0
+if '-l' in args:  # overwrite log file
+    logfile = args['-l']
+if '-d' in args:  # debug mode
+    debug = 1
 
 def parse(logline):
-    linere = re.compile('^([A-Z][a-z]{2}'') ([0-9 ]{2}) ([0-9:]{8}) \S+ kernel: \[[^\]]+\] \[([^\]]+)\] (.*)')
+    linere = re.compile('^(\d+) \S+ kernel:\[[ 0-9\.]+\]\s+DSHIELDINPUT (.*)')
     logdata = {}
     m = linere.match(logline)
     if m: 
-        month = strptime(m.group(1), '%b').tm_mon
-        if month == 12 and now.month == 1:
-            year = now.year-1
-        else:
-            year = now.year
-        day = m.group(2)
-        ltime = m.group(3)
-        date = "%s-%s-%s" % (year, month, day)
-        logtime = datetime.strptime(date+" "+ltime, '%Y-%m-%d %H:%M:%S')
-        logdata['time'] = logtime
-        if mktime(logtime.timetuple()) > startdate:
-            parts = m.group(5).split()
+        logdata['timestamp'] = datetime.fromtimestamp(int(m.group(1)))
+        logdata['time'] = logdata['timestamp'].strftime('%Y-%m-%d %H:%M:%S');
+        if m.group(1) > startdate:
+            parts = m.group(2).split()
             logdata['flags'] = ''
             for part in parts:
                 keyval = part.split('=')
                 if keyval[0] in fieldmap:
                     logdata[fieldmap[keyval[0]]] = keyval[1]
+            if logdata['dip'] == '255.255.255.255':
+                return 
+            if logdata['dip'].find('224.0.0.') == 1:
+                return
             if logdata['sip'].find(':') > 0:
                 logdata['version'] = 6
             else:
@@ -54,12 +56,14 @@ def parse(logline):
                 for fcount in tcpflagmap:
                     if fcount in parts:
                         logdata['flags'] += tcpflagmap[fcount]
-            logdata['dst'] = d.anontranslateip4(logdata['dst'])
-            logdata['src'] = d.anontranslateip4(logdata['src'])
+            logdata['dip'] = d.anontranslateip4(logdata['dip'])
+            logdata['sip'] = d.anontranslateip4(logdata['sip'])
 
             if isinstance(logdata['proto'], int):
                 return logdata
-                
+    else:
+        if debug == 1:
+            print "bad line %s" % (logline)
             
 if os.path.isfile(logfile) is None:
     sys.exit('Can not find logfile %s ' % logfile)
@@ -67,12 +71,12 @@ if os.path.isfile(logfile) is None:
 if os.path.isfile(pidfile):
     sys.exit('PID file found. Am I already running?')
 
-    
-f = open(pidfile, 'w')
-f.write(str(os.getpid()))
-f.close()
+if debug == 0:
+    f = open(pidfile, 'w')
+    f.write(str(os.getpid()))
+    f.close()
 
-if os.path.isfile(lastcount):
+if os.path.isfile(lastcount) and debug == 0:
     f = open(lastcount, 'r')
     startdate = float(f.readline())
     f.close()
@@ -91,14 +95,16 @@ with open(logfile, 'r') as f:
         data = (parse(line))
         if data is not None:
             j += 1
-            lastdate = str(mktime(data['time'].timetuple()))
-            data['time'] = data['time'].strftime('%Y-%m-%d %H:%M:%S')
+            data['time'] = data['timestamp'].strftime('%Y-%m-%d %H:%M:%S')
+            lastdate = data['timestamp'].strftime('%s')
+            del data['timestamp']
             logs.append(data)
     print json.dumps(logs)
-print "processed %d lines total and %d new lines and ended at %s" % (i, j, data['time'])
+print "processed %d lines total and %d new lines and ended at" % (i, j)
 f = open(lastcount, 'w')
 f.write(lastdate)
 f.close()
 l = {'type': 'firewall', 'logs': logs}
-d.post(l)
-os.remove(pidfile)
+if debug == 0:
+    d.post(l)
+    os.remove(pidfile)
